moved enums, renamed to BusinessDay and IInterfaces, fixed formatting and updated readme
