+++ src/main/java/customLambdaFunctions/IRepeater.java
@@ -1,0 +1,6 @@
+package customLambdaFunctions;
+
+public interface IRepeater<T> {
+
+    void repeat(int numRepeats, IPerformOperation<T> operation);
+}
