--- src/main/java/customLambdaFunctions/Repeater.java
@@ -1,6 +1,0 @@
-package customLambdaFunctions;
-
-public interface Repeater<T> {
-
-    void repeat(int numRepeats, PerformOperation<T> operation);
-}
