+++ src/main/java/genericLinkedList/PrescriptionRequestLog.java
@@ -8,15 +8,15 @@ public class PrescriptionRequestLog extends GenericQueue<Prescription> {
 
     private static final Logger LOG = LogManager.getLogger(PrescriptionRequestLog.class);
 
-    public void addPrescriptionRequest(Prescription prescription){
+    public void addPrescriptionRequest(Prescription prescription) {
         enqueue(prescription);
     }
 
-    public Prescription getNextPrescriptionRequest(){
+    public Prescription getNextPrescriptionRequest() {
         if (isEmpty()) {
-        LOG.warn("The prescriptionRequestLog is empty");
-        return null;
-    }
+            LOG.warn("The prescriptionRequestLog is empty");
+            return null;
+        }
         return dequeue();
     }
