+++ src/main/java/product/Medication.java
@@ -1,5 +1,6 @@
 package product;
 
+import enums.ProductType;
 import java.util.Objects;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
